updating frameworks and build scripts
