+++ script/version.json
@@ -1,0 +1,8 @@
+{
+  "inherit": true,
+  "cloudBuild": {
+    "buildNumber": {
+      "enabled": true
+    }
+  }
+}
