+++ src/Akavache.Tests/Akavache.Tests.csproj
@@ -1,25 +1,27 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net452</TargetFrameworks>
+    <TargetFrameworks>net461</TargetFrameworks>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <GenerateBindingRedirectsOutputType>true</GenerateBindingRedirectsOutputType>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
     <PackageReference Include="Microsoft.Reactive.Testing" Version="[3.1.1,4)" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="xunit" Version="2.2.0" />
-    <PackageReference Include="xunit.runner.console" Version="2.2.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
-    <PackageReference Include="Splat" Version="2.0.0" />
-    <PackageReference Include="reactiveui" Version="8.0.0-alpha0034" />
-    <PackageReference Include="reactiveui-testing" Version="8.0.0-alpha0034" />
-    <PackageReference Include="SQLitePCLRaw.lib.e_sqlite3.v110_xp" Version="1.1.8" /> 
+		<PackageReference Include="xunit" Version="2.4.0-beta.2.build3984" />
+		<PackageReference Include="xunit.runner.console" Version="2.4.0-beta.2.build3984" />
+		<PackageReference Include="xunit.runner.visualstudio" Version="2.4.0-beta.2.build3984" />
+		<PackageReference Include="Xunit.StaFact" Version="0.2.9" />
+    <PackageReference Include="reactiveui" Version="8.2.1" />
+    <PackageReference Include="ReactiveUI.Testing" Version="8.2.1" />
+    <PackageReference Include="SQLitePCLRaw.lib.e_sqlite3.v110_xp" Version="1.1.11" />
+		<PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
+		<PackageReference Include="System.Reactive" Version="[3.1.1,4)" />
+		<PackageReference Include="Splat" Version="4.0.0" />
   </ItemGroup>
 
   <ItemGroup>
