+++ src/Directory.build.props
@@ -16,12 +16,21 @@
     <DebugType>full</DebugType>
     <IsTestProject>$(MSBuildProjectName.Contains('Tests'))</IsTestProject>
   </PropertyGroup>
-  
-  <ItemGroup>
-    <PackageReference Include="MSBuild.Sdk.Extras" Version="1.0.6" PrivateAssets="All" />
-  </ItemGroup>
 
-  <ItemGroup Condition="'$(IsTestProject)' != 'true' and '$(SourceLinkEnabled)' != 'false'">
-    <PackageReference Include="SourceLink.Create.CommandLine" Version="2.1.2" PrivateAssets="All" /> 
-  </ItemGroup>
+	<!-- Use a Choose / When so the legacy project system doesn't get this -->
+	<Choose>
+		<When Condition="'$(IsLegacyProject)' != 'true'">
+			<ItemGroup>
+				<PackageReference Include="MSBuild.Sdk.Extras" Version="1.3.1" PrivateAssets="All" />
+			</ItemGroup>
+		</When>
+	</Choose>
+
+	<ItemGroup>
+		<PackageReference Include="Nerdbank.GitVersioning" Version="2.1.23" PrivateAssets="all" />
+	</ItemGroup>
+
+	<ItemGroup Condition="'$(IsTestProject)' != 'true' and '$(SourceLinkEnabled)' != 'false'">
+		<PackageReference Include="SourceLink.Create.CommandLine" Version="2.8.0" PrivateAssets="All" />
+	</ItemGroup>
 </Project>
